# src/backend/TestPlanService/Api/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "127.0.0.1:5000",
  "schemes": [
    "http"
  ],
  "paths": {
    "/projects/{projectId}/plans": {
      "get": {
        "tags": [
          "Plans"
        ],
        "operationId": "Plans_GetPlans",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetPlansResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Plans"
        ],
        "operationId": "Plans_AddPlan",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddOrUpdatePlanRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
    "/projects/{projectId}/plans/{planId}": {
      "put": {
        "tags": [
          "Plans"
        ],
        "operationId": "Plans_UpdatePlan",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "planId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddOrUpdatePlanRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Plans"
        ],
        "operationId": "Plans_DeletePlan",
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "planId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/projects/{projectId}/plans/{planId}/dublicate": {
      "post": {
        "tags": [
          "Plans"
        ],
        "operationId": "Plans_DublicatePlan",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "planId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
    "/projects/{projectId}/plan/{planId}": {
      "get": {
        "tags": [
          "Plans"
        ],
        "operationId": "Plans_GetPlanInfo",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "planId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/TestPlanItem"
            }
          }
        }
      }
    },
...
    "/projects/{projectId}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_DeleteProject",
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TestSuites"
        ],
        "operationId": "TestSuites_MoveSuite",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "id",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "moveSuiteRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MoveSuiteRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "GetPlansResponse": {
      "type": "object",
      "properties": {
        "plans": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TestPlanItem"
          }
        }
      }
    },
    "TestPlanItem": {
      "type": "object",
      "required": [
        "id",
        "state"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "state": {
          "$ref": "#/definitions/PlanState"
        },
        "assignedTo": {
          "$ref": "#/definitions/UserIdentity"
        }
      }
    },
    "PlanState": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Planed",
        "Active",
        "Closed"
      ],
      "enum": [
        0,
        1,
        2
      ]
    },
    "UserIdentity": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "AddOrUpdatePlanRequest": {
      "type": "object",
      "required": [
        "state",
        "assignedTo"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "state": {
          "$ref": "#/definitions/PlanState"
        },
        "assignedTo": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "MoveSuiteRequest": {
      "type": "object",
      "required": [
        "newParentId"
      ],
      "properties": {
        "newParentId": {
          "type": "integer",
          "format": "int32"
        }
      }
    }
  }
}